Refactor code into a separate AttributeListBuilder
