+++ src/app/server-cluster/BUILD.gn
@@ -15,6 +15,8 @@ import("//build_overrides/chip.gni")
 
 source_set("server-cluster") {
   sources = [
+    "AttributeListBuilder.cpp",
+    "AttributeListBuilder.h",
     "DefaultServerCluster.cpp",
     "DefaultServerCluster.h",
     "ServerClusterContext.h",
